feat: decompose hooks context class into functions
